fix QbertLevel rotation
Turns out it was very simple to perform the rotation we wanted - the trick was to use the transform.basis property. When we call its rotated method, we get to pick the vector we want to rotate about, and it solves things from there.

The axis of rotation that we wanted was pointing straight at the camera - or, more precisely, it was the vector where x, y, and z were all equal. Hence Vector3(1,1,1).normalized()

It's important that we do this operation inside the GridMap instead of the CubeFaces scene - we want to rotate the whole structure, not the individual pieces inside of it.

One side note - be wary of integer division when specifying how much the whole thing should rotate. 2 / 3 * PI (which seemed more readable to me) does integer division with 2 / 3, which truncates to 0. Instead, you can do what I did here (multiply PI in at the start to make it a floating point number) or you could do 2.0 / 3.0 * PI.